+++ CHANGELOG.md
@@ -31,6 +31,21 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   (#855)
 - Configure TLS using the system CA for OTLP gRPC exporter connections when
   configured to connect to external endpoints. (#792)
+  
+### Removed
+
+- Minimize `github.com/signalfx/splunk-otel-go/distro` API to
+  contain only necessary option functions. (#941)
+  - Remove `WithAccessToken` function,
+    use `SPLUNK_ACCESS_TOKEN` environment variable instead.
+  - Remove `WithEndpoint` function,
+    use one of the
+    `OTEL_EXPORTER_OTLP_ENDPOINT`, `OTEL_EXPORTER_OTLP_TRACES_ENDPOINT`, `OTEL_EXPORTER_JAEGER_ENDPOINT`
+    environment variables instead.
+  - Remove `WithPropagator` function,
+    use `OTEL_PROPAGATORS` environment variable instead.
+  - Remove `WithTraceExporter` function,
+    use `OTEL_TRACES_EXPORTER` environment variable instead.
 
 ### Fixed